context sensative parsing of unquote in quasiquote
This seems like a good idea, and xemacs actually implements it,
unfortunately it summons the outsiders from the turing complete void
to eat your sanity and spread chaos and complexity all over everything
and creats major problems if you don't also make the unquote symbol
whitespace when not in quasiquote which again, very confusing, and to
avoid it you have to add even more complexity so that `(a ,b,c) will
read as (list 'a b c) instead of as (list 'a b,c). Other possible
readings are (list 'a b 'c) if comma is whitespace outside quasi.

In most traditional lisps you can just escape the comman if you really
need it in a symbol since they have almost completely free-form
identifiers, it seems vastly more sensible. I'm implementing it anyway
so that it is avaiable and so that all the tradeoffs of doingso can be
visible in the event that one of them might be worth the tradeoff.